'LOGIN' and 'REGISTRATION' ready to use
When starting the client, it is suggested to enter login and password. If the user is already registered, then the validity of the password and the existence of this user online are checked. If the entered name is not in the database, then the user will receive an account in the database. There is a bug with changing rooms.